pyproject.toml
[tool.poetry]
name = "module8"
version = "0.1.0"
description = ""
authors = ["Nastya Sol <an5oldatova@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
djangorestframework = "^3.15.2"
load-dotenv = "^0.1.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
coverage = "^7.6.12"
celery = "^5.4.0"
drf-yasg = "^1.21.10"
stripe = "^11.6.0"
redis = "^5.2.1"
django-celery-beat = "^2.7.0"


[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
